converter/requirements.txt
dnspython==2.2.1
pika==1.3.1
Pillow==9.4.0
pymongo==4.3.3
python-dateutil==2.8.2
six==1.16.0
